Added better color, indicators, and a README
Each indicator ('+' or '-') means that the price has been above (+) or
below (-) the 24hr average for an hour. So, [BTC++] means that the
current price has been above the average price for 2 hours.

The BTC bar element will be red when current price is below average and
green while above average.

The wrapper script also removes any unused elements by removing any
full_text with 'no' or 'down' in it....probably a better way, but I got
tired of looking at no IPv6 in red.